refactor(RoomView): colocate hook param and result types with their hooks
